Asynchronous power handling.
Replace a static, single thread for System and Device power transitions with
an IO_WORKITEM for each transition. IO_WORKITEMs are only used when a transition
requires calling codepaths that must be called at PASSIVE_LEVEL.
Move the bulk of power transitions to pending IRPs and IRP completion routines.

Signed-off-by: Martin Harvey <martin.harvey@citrix.com>

Refactored to only apply to XenBus power

Signed-off-by: Owen Smith <owen.smith@cloud.com>

Cleaned up out-of-style comments.

Signed-off-by: Paul Durrant <paul@xen.org>