Deprecate PDO revisions 0x09000007 and earlier
0x09000008 has been available since 2021.

Deprecate the old PDO revisions to make space for new revisions.

Due to b0b9cc139600 removing REV from DeviceID, it's no longer safe to
remove interfaces even if they're not exposed by the revision list, as
old drivers could have bound to older revisions and not aware of updated
Xenbus.

The following interface versions are deprecated but not removed:
* SHARED_INFO v2
* EVTCHN v5-v8
* CACHE v1
* GNTTAB v2-v3
* EMULATED v1

Signed-off-by: Tu Dinh <ngoc-tu.dinh@vates.tech>