Merge all hardware IDs into one Models entry
This makes sure that on driver installation, all supported hardware IDs
will be populated into the driver database, thus allowing the device ID
to be freely switched without causing boot installation issues.

Bump the feature score to compensate for the resulting degradation of
identifier-match-type score.

Note that if the vendor device is disabled, merging hardware IDs will
cause the 0002 device to have a higher identifier score than 0001.
Still, omitting the hardware ID in the Models entry as documented in MS
docs will cause an Inf2Cat warning "22.9.3: Missing hardware ID (cannot
have just compatible ID)", so we just have to deal with the score
promotion in this case.

Signed-off-by: Tu Dinh <ngoc-tu.dinh@vates.tech>